+++ src/core/activate.ts
@@ -1,18 +1,19 @@
 import { loadTranslations, getVscodeLang } from 'vscode-ext-localisation';
 import * as vscode from 'vscode';
-import { PackageList } from '../treeviews';
+import { FolderList, PackageList } from '../treeviews';
 import { registerCommands } from '../commands';
 import { registerWebviews } from '../webviews';
 import { setupSidebar } from '../sidebar';
 
-let packageJsonDataProvider: PackageList | undefined = undefined;
-
 export const setupExt = (context: vscode.ExtensionContext, lang: string): void => {
   loadTranslations(lang, context.extensionPath);
-  packageJsonDataProvider = new PackageList(context);
-  registerCommands(context, packageJsonDataProvider);
+
+  const packageJsonDataProvider = new PackageList(context);
+  const folderTreeDataProvider = new FolderList(vscode.workspace.workspaceFolders, context);
+
+  registerCommands(context, folderTreeDataProvider, packageJsonDataProvider);
   registerWebviews(context);
-  setupSidebar(context, vscode.workspace.workspaceFolders, packageJsonDataProvider);
+  setupSidebar(context, folderTreeDataProvider, packageJsonDataProvider);
 };
 
 export const activate = (context: vscode.ExtensionContext): void => {